Kernel/Net: Move Realtek network adapters code to a separate directory
