+++ Kernel/CMakeLists.txt
@@ -168,14 +168,14 @@ set(KERNEL_SOURCES
     Net/Intel/E1000ENetworkAdapter.cpp
     Net/Intel/E1000NetworkAdapter.cpp
     Net/NE2000/NetworkAdapter.cpp
+    Net/Realtek/RTL8139NetworkAdapter.cpp
+    Net/Realtek/RTL8168NetworkAdapter.cpp
     Net/IPv4Socket.cpp
     Net/LocalSocket.cpp
     Net/LoopbackAdapter.cpp
     Net/NetworkAdapter.cpp
     Net/NetworkTask.cpp
     Net/NetworkingManagement.cpp
-    Net/RTL8139NetworkAdapter.cpp
-    Net/RTL8168NetworkAdapter.cpp
     Net/Routing.cpp
     Net/Socket.cpp
     Net/TCPSocket.cpp
