+++ Kernel/Net/Realtek/RTL8168NetworkAdapter.cpp
@@ -9,7 +9,7 @@
 #include <Kernel/Bus/PCI/IDs.h>
 #include <Kernel/Debug.h>
 #include <Kernel/Net/NetworkingManagement.h>
-#include <Kernel/Net/RTL8168NetworkAdapter.h>
+#include <Kernel/Net/Realtek/RTL8168NetworkAdapter.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
